Release 3.0.0

+++ src/pages/Register/index.js
@@ -1,10 +1,8 @@
 import React, { useState } from 'react';
-
 import api from '../../services/api';
 
-const Login = ({ history }) => {
-  const [errors, setErrors] = useState();
-
+const Register = ({ history }) => {
+  const [errors, setErrors] = useState(null);
   const containerStyle = {
     display: 'flex',
     justifyContent: 'center',
@@ -17,25 +15,40 @@ const Login = ({ history }) => {
     backgroundSize: 'cover',
   };
 
-  const handleSubmit = event => {
+  const handleSubmit = (event) => {
     event.preventDefault();
     const data = new FormData(event.target);
 
+    const password = data.get('password');
+    const passwordConfirm = data.get('password-confirm');
+    // Validation
+    if (password && passwordConfirm) {
+      if (password !== passwordConfirm) {
+        alert('As senhas digitadas não conferem');
+        return;
+      }
+      if (password.length < 6) {
+        alert('A senha deve possuir ao menos 6 caracteres');
+        return;
+      }
+    }
+
     api
-      .post('/authenticate', {
+      .post('/register', {
+        name: data.get('name'),
         email: data.get('email'),
         password: data.get('password'),
+        passwordConfirm: data.get('password-confirm'),
       })
-      .then(res => {
+      .then((res) => {
         const { user, token } = res.data;
         sessionStorage.setItem('@kan/currentuser', user);
         sessionStorage.setItem('@kan/token', token);
         history.push('/boards');
-        return;
       })
       .catch((err) => {
         setErrors(err.response.data.error);
-        console.log(err);
+        alert('Ocorreu um erro. Preencha todos os campos corretamente.');
       });
   };
 
@@ -44,7 +57,27 @@ const Login = ({ history }) => {
       <div className="card" style={{ width: '70vw' }}>
         <form onSubmit={handleSubmit} method="POST">
           <div className="card-body">
-            <h1 className="card-title">Login</h1>
+            <h1 className="card-title">Crie sua conta</h1>
+            {
+              errors
+                ? (
+                  <div className="alert alert-danger">
+                    {errors.map((error, index) => <p key={index}>{error}</p>)}
+                  </div>
+                )
+                : null
+            }
+            <div className="form-group">
+              <label htmlFor="email">Nome</label>
+              <input
+                id="name"
+                name="name"
+                type="text"
+                className="form-control"
+                minLength={1}
+                required
+              />
+            </div>
             <div className="form-group">
               <label htmlFor="email">E-mail</label>
               <input
@@ -65,14 +98,20 @@ const Login = ({ history }) => {
                 required
               />
             </div>
-            <p>{ errors }</p>
-            <a href="/register" className="btn">
-              Registrar-me
-            </a>
+            <div className="form-group">
+              <label htmlFor="password">Confime sua senha</label>
+              <input
+                id="password-confirm"
+                name="password-confirm"
+                type="password"
+                className="form-control"
+                required
+              />
+            </div>
           </div>
           <div className="card-footer">
             <button type="submit" className="btn btn-dark btn-lg">
-              Entrar
+              Registrar
             </button>
           </div>
         </form>
@@ -81,4 +120,4 @@ const Login = ({ history }) => {
   );
 };
 
-export default Login;
+export default Register;
